Lots of changes
Moved the tag and share dialogs to lib unicorn so that the radio player
can open them. It now opens them as well.

Added QtSingleApplication so that all unicorn application dervive from
this.  We can now only have one version of the radio or the scrobbler at
a time.

Opening the radio app if it's already open will pass any arguments on to
the already running app. This means we can start radio tracks from the
website in the already open radio.