Lots more changes for the combo-station ui.
It looks rather nice now although there are still a lot of things that
can be done.

Multi-user can select intersection or union.

Bar charts behind the top artist and top tag lists. We cannot get the
tastometer results for users so have no bar cahrt fro this.

Icons are displayed for list items, but we need a default user and a
default artist icon.

The UI is not very smooth when you switch to the combo station starter.

Drag and Drop now works.